Replaced the original dGroup implementation with the latest (previosuly included in dGroup2)
